post

+++ app/Http/Controllers/TasksController.php
@@ -15,15 +15,23 @@ class TasksController extends Controller
      */
     public function index()
     {
-        // task一覧を取得
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        // task一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
-
+    
     // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
     public function create()
     {
@@ -39,16 +47,17 @@ class TasksController extends Controller
     public function store(StoreBlogPost $request)
     {   $validated = $request->validated();
     
-        //dd($validated); 
-        
-        // taskを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        // 認証済みユーザ（閲覧者）タスク稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+             'status' => $request->status
+            
+        ]);
+
 
         // トップページへリダイレクトさせる
         return redirect('/');
+
     }
 
     // getでtask/idにアクセスされた場合の「取得表示処理」
@@ -56,22 +65,27 @@ class TasksController extends Controller
     {
         // idの値でtaskを検索して取得
         $task = Task::findOrFail($id);
+        
 
         // task詳細ビューでそれを表示
-        return view('tasks.show', [
+        if (\Auth::id() === $task->user_id){
+            return view('tasks.show', [
             'task' => $task,
         ]);
+        }
     }
 
      public function edit($id)
     {
         // idの値でtaskを検索して取得
         $task = Task::findOrFail($id);
-
+        
         // task編集ビューでそれを表示
-        return view('tasks.edit', [
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
             'task' => $task,
         ]);
+        }
     }
 
        public function update(StoreBlogPost $request, $id)
@@ -81,10 +95,12 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         // タスクを更新
+        
+        if (\Auth::id() === $task->user_id){
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
-        
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -94,8 +110,11 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
+        
+        // 認証済みユーザ（閲覧者）がその侑の所有者である場合は、侑を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
